# Pool zstd encoding/decoding