--- a/tempodb/backend/tenantindex_benchmark_test.go
+++ b/tempodb/backend/tenantindex_benchmark_test.go
@@ -85,3 +85,44 @@ func BenchmarkIndexUnmarshal(b *testing.B) {
 		_ = unIdx.unmarshal(buf)
 	}
 }
+
+func BenchmarkIndexUnmarshalPb(b *testing.B) {
+	idx := &TenantIndex{
+		Meta: []*BlockMeta{
+			NewBlockMeta("test", uuid.New(), "v1", EncGZIP, "adsf"),
+			NewBlockMeta("test", uuid.New(), "v2", EncNone, "adsf"),
+			NewBlockMeta("test", uuid.New(), "v3", EncLZ4_4M, "adsf"),
+		},
+		CompactedMeta: []*CompactedBlockMeta{
+			{
+				BlockMeta:     *NewBlockMeta("test", uuid.New(), "v1", EncGZIP, "adsf"),
+				CompactedTime: time.Now(),
+			},
+			{
+				BlockMeta:     *NewBlockMeta("test", uuid.New(), "v1", EncZstd, "adsf"),
+				CompactedTime: time.Now(),
+			},
+			{
+				BlockMeta:     *NewBlockMeta("test", uuid.New(), "v1", EncSnappy, "adsf"),
+				CompactedTime: time.Now(),
+			},
+		},
+	}
+
+	for i := range idx.Meta {
+		idx.Meta[i].DedicatedColumns = DedicatedColumns{
+			{Scope: "resource", Name: "namespace", Type: "string"},
+			{Scope: "span", Name: "http.method", Type: "string"},
+			{Scope: "span", Name: "namespace", Type: "string"},
+		}
+	}
+
+	buf, err := idx.marshalPb()
+	require.NoError(b, err)
+
+	unIdx := &TenantIndex{}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = unIdx.unmarshalPb(buf)
+	}
+}