@zalegrala 2024-10-18
```
❯ benchstat main.txt zstdPooling.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/backend
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
                    │   main.txt   │           zstdPooling.txt           │
                    │    sec/op    │   sec/op     vs base                │
IndexUnmarshalPb-16   20.645µ ± 1%   3.694µ ± 2%  -82.11% (p=0.000 n=10)

                    │    main.txt    │           zstdPooling.txt            │
                    │      B/op      │     B/op      vs base                │
IndexUnmarshalPb-16   154.434Ki ± 0%   3.709Ki ± 1%  -97.60% (p=0.000 n=10)

                    │  main.txt   │          zstdPooling.txt           │
                    │  allocs/op  │ allocs/op   vs base                │
IndexUnmarshalPb-16   100.00 ± 0%   59.00 ± 0%  -41.00% (p=0.000 n=10)
```